Modernization. Remove outdated tiffutil calls.

+++ inform/Project/Syntax/IFInform6Highlighter.m
@@ -682,7 +682,7 @@ static int compare(const void* a, const void* b) {
 	return [IFProjectPane attributeForStyle: style];
 }
 
-- (float) tabStopWidth {
+- (CGFloat) tabStopWidth {
 	return 28.0;
 }
